requirements.txt
asgiref==3.3.4
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2
django-cors-headers==3.7.0
django-extensions==3.1.3
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
drf-yasg==1.20.0
idna==2.10
iniconfig==1.1.1
itypes==1.2.0
Jinja2==2.11.3
lazy-object-proxy==1.6.0
MarkupSafe==1.1.1
model-bakery==1.3.1
packaging==20.9
pluggy==0.13.1
pre-commit==2.12.1
psycopg2==2.8.6
py==1.10.0
PyJWT==1.7.1
pyparsing==2.4.7
pytest==6.2.3
pytest-drf==1.1.2
pytest-django==4.2.0
pytest-fixture-order==0.1.3
pytest-lambda==1.2.4
python-dotenv==0.17.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.4
Werkzeug==2.0.0
wrapt==1.12.1
